refactor tests for topography module
